Complete frontend interactions after booking a slot
