+++ src/backend/bookingRoutes.ts
@@ -1,7 +1,12 @@
 import express from 'express';
 
-import { createBooking } from './bookingController';
-import { UserRole, ApiBooking, ApiCreateBookingParams } from '../common/constants-common';
+import { createBooking, getAllBookings } from './bookingController';
+import {
+  UserRole,
+  ApiBooking,
+  ApiCreateBookingParams,
+  ApiBookedSlot,
+} from '../common/constants-common';
 import { isAuthenticated } from './middlewares';
 
 const router = express.Router();
@@ -20,7 +25,7 @@ router.post<
   }
 
   // Only allow staff members to create bookings for any other users
-  if (!!req.body.userUuid && user.uuid && user.role !== UserRole.staff) {
+  if (req.body.userUuid !== user.uuid && user.role !== UserRole.staff) {
     res.status(401).json({ error: 'unauthorized' });
     return;
   }
@@ -42,6 +47,45 @@ router.post<
   }
 
   res.status(201).json({ data: newBooking });
+});
+
+router.get<
+  Record<string, never>,
+  { data: Array<ApiBookedSlot> } | { error: string },
+  Record<string, never>,
+  { startDate: string; endDate: string }
+>('/', isAuthenticated([UserRole.staff, UserRole.volunteer]), async (req, res) => {
+  const allBookings = await getAllBookings();
+  if (allBookings === null) {
+    res.status(400).json({ error: 'unable to get booking data' });
+    return;
+  }
+
+  if (!req.query?.startDate || !req.query?.endDate) {
+    res.status(400).json({ error: 'Missing startDate or endDate in request body' });
+    return;
+  }
+
+  const startDate = new Date(req.query.startDate);
+  const endDate = new Date(req.query.endDate);
+
+  const bookedSlotsCount = allBookings
+    .filter(({ start, end }) => startDate <= new Date(start) && endDate >= new Date(end))
+    .reduce((result, { start, end, bookingType }) => {
+      const existingSlotIndex = result.findIndex(
+        (r) => r.start === start && r.end === end && r.bookingTypeUuid === bookingType.uuid,
+      );
+      if (existingSlotIndex < 0) {
+        return [...result, { start, end, bookingTypeUuid: bookingType.uuid, count: 1 }];
+      } else {
+        result[existingSlotIndex] = {
+          ...result[existingSlotIndex],
+          count: result[existingSlotIndex].count + 1,
+        };
+        return result;
+      }
+    }, [] as Array<ApiBookedSlot>);
+  res.status(200).json({ data: bookedSlotsCount });
 });
 
 export default router;
