+++ src/frontend/Volunteer.tsx
@@ -5,6 +5,7 @@ import { UserRole } from '../common/constants-common';
 import { useAuth } from './AuthContext';
 import { Letters } from './Letters';
 import { Reply } from './Reply';
+import { Booking } from './pages/Booking';
 
 export const Volunteer: React.FunctionComponent<RouteComponentProps> = () => {
   const { token, user } = useAuth();
@@ -13,6 +14,7 @@ export const Volunteer: React.FunctionComponent<RouteComponentProps> = () => {
     <Redirect noThrow to="/" />
   ) : (
     <Router>
+      <Booking path="booking" />
       <Letters path="letters" />
       <Reply path="letters/:letterUuid" />
     </Router>
